Remove incorrect delay from ConnectAndPublish()
Without this change the absolute maximum number of received messages
per second is 100.  With this change, I can get the full 600 msg/s
from my CAN bus and the CPU usage is lower.

Reading from "can0" on linux is blocking by default with this package
so things Just Work in proper Go fashion without the sleep.